Updated serialization and effects manager type.
