+++ SadConsole.Core/Consoles/TextSurface.cs
@@ -8,262 +8,6 @@ using System.Text;
 
 namespace SadConsole.Consoles
 {
-    public class EffectsManager
-    {
-        [DataMember(Name = "Effects")]
-        protected Dictionary<ICellEffect, int[]> _effectSerialized;
-        protected Dictionary<ICellEffect, CellEffectData> _effects;
-        protected Dictionary<Cell, CellEffectData> _effectCells;
-
-        protected ITextSurface backingSurface;
-
-        public EffectsManager(ITextSurface surface)
-        {
-            _effects = new Dictionary<ICellEffect, CellEffectData>(20);
-            _effectCells = new Dictionary<Cell, CellEffectData>(50);
-            backingSurface = surface;
-        }
-
-        public void Hydrate(ITextSurface surface)
-        {
-            if (backingSurface != null)
-                throw new Exception("Effects manager was created with an existing surface. This method is for serialization loading.");
-
-            backingSurface = surface;
-
-            _effects = new Dictionary<ICellEffect, CellEffectData>(20);
-            _effectCells = new Dictionary<Cell, CellEffectData>(50);
-
-            if (_effectSerialized != null)
-            {
-                foreach (var effect in _effectSerialized.Keys)
-                {
-                    int[] effectCellIndexes = _effectSerialized[effect];
-
-                    List<Cell> cells = new List<Cell>(effectCellIndexes.Length);
-
-                    foreach (var index in effectCellIndexes)
-                        cells.Add(backingSurface.Cells[index]);
-
-                    var effectData = new CellEffectData(effect) { Cells = cells };
-
-                    _effects.Add(effect, effectData);
-
-                    foreach (var cell in cells)
-                        _effectCells.Add(cell, effectData);
-                }
-            }
-
-            _effectSerialized = null;
-        }
-
-        /// <summary>
-        /// Changes the effect of a specific cell.
-        /// </summary>
-        /// <param name="cell">Cells to change the effect on.</param>
-        /// <param name="effect">The effect to associate with the cell.</param>
-        public void SetEffect(Cell cell, ICellEffect effect)
-        {
-            var list = new List<Cell>(backingSurface.Cells);
-            if (!list.Contains(cell))
-                throw new Exception("Cell is not part of the surface used to create this effects manager.");
-
-            if (effect != null)
-            {
-                CellEffectData workingEffect;
-
-                if (effect.CloneOnApply)
-                {
-                    effect = effect.Clone();
-                    workingEffect = new CellEffectData(effect);
-                    _effects.Add(workingEffect.Effect, workingEffect);
-                }
-                else
-                {
-                    // Is the effect unknown? Add it.
-                    if (GetKnownEffect(effect, out workingEffect) == false)
-                        _effects.Add(workingEffect.Effect, workingEffect);
-                    else
-                        if (workingEffect.Cells.Contains(cell))
-                        return;
-                }
-
-                // Remove the targeted cell from the known cells list if it is already there (associated with another effect)
-                ClearCellEffect(cell);
-
-                // Add the cell to the effects by cell key and to list of known cells for the effect
-                _effectCells.Add(cell, workingEffect);
-                cell.Effect = workingEffect.Effect;
-                workingEffect.Cells.Add(cell);
-            }
-            else
-                ClearCellEffect(cell);
-        }
-
-        /// <summary>
-        /// Changes the effect of the <paramref name="cells"/> provided.
-        /// </summary>
-        /// <param name="cells">Cells to change the effect on.</param>
-        /// <param name="effect">The effect to associate with the cell.</param>
-        public void SetEffect(IEnumerable<Cell> cells, ICellEffect effect)
-        {
-            var list = new List<Cell>(backingSurface.Cells);
-            foreach (var cell in cells)
-            {
-                if (!list.Contains(cell))
-                    throw new Exception("Cell is not part of the surface used to create this effects manager.");
-            }
-
-
-            if (effect != null)
-            {
-                CellEffectData workingEffect;
-
-                if (effect.CloneOnApply)
-                {
-                    effect = effect.Clone();
-                    workingEffect = new CellEffectData(effect);
-                    _effects.Add(workingEffect.Effect, workingEffect);
-                }
-                else
-                {
-                    // Is the effect unknown? Add it.
-                    if (GetKnownEffect(effect, out workingEffect) == false)
-                        _effects.Add(workingEffect.Effect, workingEffect);
-                }
-
-                foreach (var cell in cells)
-                {
-                    if (!workingEffect.Cells.Contains(cell))
-                    {
-                        // Remove the targeted cell from the known cells list if it is already there (associated with another effect)
-                        ClearCellEffect(cell);
-
-                        // Add the cell to the effects by cell key and to list of known cells for the effect
-                        _effectCells.Add(cell, workingEffect);
-                        cell.Effect = workingEffect.Effect;
-                        workingEffect.Cells.Add(cell);
-                    }
-                }
-            }
-            else
-            {
-                foreach (var cell in cells)
-                    ClearCellEffect(cell);
-            }
-        }
-
-        /// <summary>
-        /// Gets the effect of the specified cell.
-        /// </summary>
-        /// <param name="x">The x location of the cell.</param>
-        /// <param name="y">The y location of the cell.</param>
-        /// <returns>The effect.</returns>
-        public Effects.ICellEffect GetEffect(Cell cell)
-        {
-            return _effectCells.ContainsKey(cell) ? _effectCells[cell].Effect : null;
-        }
-
-
-        #region Effect Helpers
-        protected bool GetKnownEffect(ICellEffect effect, out CellEffectData effectData)
-        {
-            if (_effects.TryGetValue(effect, out effectData))
-            {
-                return true;
-            }
-            else
-            {
-                effectData = new CellEffectData(effect);
-                return false;
-            }
-        }
-
-        protected void ClearCellEffect(Cell cell)
-        {
-            CellEffectData oldEffectData;
-
-            if (_effectCells.TryGetValue(cell, out oldEffectData))
-            {
-                oldEffectData.Effect.Clear(cell);
-                oldEffectData.Cells.Remove(cell);
-                _effectCells.Remove(cell);
-
-                if (oldEffectData.Cells.Count == 0)
-                    _effects.Remove(oldEffectData.Effect);
-            }
-
-            cell.Effect = null;
-        }
-
-        /// <summary>
-        /// Updates all known effects and applies them to their associated cells.
-        /// </summary>
-        /// <param name="timeElapsed">The time elapased since the last update.</param>
-        public void UpdateEffects(double timeElapsed)
-        {
-            List<ICellEffect> effectsToRemove = new List<ICellEffect>();
-
-            foreach (var effectData in _effects.Values)
-            {
-                List<Cell> cellsToRemove = new List<Cell>();
-                effectData.Effect.Update(timeElapsed);
-
-                foreach (var cell in effectData.Cells)
-                {
-                    effectData.Effect.Apply(cell);
-
-                    if (effectData.Effect.IsFinished && effectData.Effect.RemoveOnFinished)
-                        cellsToRemove.Add(cell);
-                }
-
-                foreach (var cell in cellsToRemove)
-                {
-                    effectData.Effect.Clear(cell);
-                    effectData.Cells.Remove(cell);
-                    _effectCells.Remove(cell);
-                    cell.Effect = null;
-                }
-
-                if (effectData.Cells.Count == 0)
-                    effectsToRemove.Add(effectData.Effect);
-            }
-
-            foreach (var effect in effectsToRemove)
-                _effects.Remove(effect);
-        }
-        #endregion
-
-        [OnSerializing]
-        private void BeforeSerialized(StreamingContext context)
-        {
-            _effectSerialized = new Dictionary<ICellEffect, int[]>(_effects.Count);
-            List<Cell> currentCells = new List<Cell>(backingSurface.Cells);
-
-            foreach (var effectData in _effects.Values)
-            {
-                List<int> effectCellPositions = new List<int>(effectData.Cells.Count);
-
-                foreach (var cell in effectData.Cells)
-                    effectCellPositions.Add(currentCells.IndexOf(cell));
-
-                _effectSerialized.Add(effectData.Effect, effectCellPositions.ToArray());
-            }
-        }
-
-        protected class CellEffectData
-        {
-            public ICellEffect Effect;
-            public List<Cell> Cells;
-
-            public CellEffectData(ICellEffect effect)
-            {
-                Effect = effect;
-                Cells = new List<Cell>();
-            }
-        }
-    }
-
     /// <summary>
     /// Represents all the basic information about console text and methods to manipulate that text.
     /// </summary>
@@ -1440,18 +1184,27 @@ namespace SadConsole.Consoles
         #endregion
 
         
+        /// <summary>
+        /// Saves this TextSurface.
+        /// </summary>
+        /// <param name="file">The file to save the TextSurface too.</param>
         public void Save(string file)
         {
             TextSurfaceSerialized.Save(this, file);
         }
 
+        /// <summary>
+        /// Loads a TextSurface.
+        /// </summary>
+        /// <param name="file">The file to load.</param>
+        /// <returns></returns>
         public static TextSurface Load(string file)
         {
             return TextSurfaceSerialized.Load(file);
         }
 
         [DataContract]
-        public class TextSurfaceSerialized
+        internal class TextSurfaceSerialized
         {
             [DataMember]
             Cell[] Cells;
